Fixed last_value counting for in_position
Previously last_value would only work properly if you were evaluating the top lines of the file so it would be .../1/*, /2/, /3/ etc.  rather than some lines mid way down that would be .../17/*, /18/, /19/ etc.  In siblings we actually need to count items, so we now convert last() to be the count; minus 1 if there's more than one object, otherwise it should stay as 1.  Also, we should return true if the result of last_value is 1 or less(if less it means something probably went wrong because it didn't return any values, so likely that will never happen, but if 1... then it means we have no items of that type or xpath to compare to anyways, so it will never be, before or after, itself; therefore, always in_position.